Fixes on Vortigern
By rotating the point and find new forward vector, it seems to work out
well. However as noted, if you want to use more than 9 projectiles, you
have to use other units such as thinker or dummy to store the projectile
instead, but shouldn't be much more work if it's really necessary.