More bugfixes
Fix for astolfo refunding too much mana when attempting R in marbles or
when 2 raids are present.
Fix for tamamo's D disappearing for the entire game when
TamamoCheckCombo() swaps the abilities and the D attribute is added.
Fix for lancelot E-R abuse and being able to use KoH skills during
arondite. I believe it should (and seems to) work fine but probably
needs more testing against edge cases.