Fixes to Lancer5 and Lancelot
Will avoid touching issue#80 since nothing about tooltips suggest BC
cooldown to be decreased from 60 to 30 when attribute is taken.

Lancelot:
For Issue#83, fix is at line 226 to 238 where SwapAbilities(,,true,)
becomes SwapAbilities(,,caster.KnightLevel~=nil,)
For Issue#56, fix is at line 230 and also 617 and 625
For Issue#55, fix is at line 171, 180 and beyond. As a result, line 317
to 325 may no longer be necessary since it now appears to me that it is
now impossible to get into spellbook with arondite activated. I could be
wrong so I will leave it there for now.

Lancer5:
For line 47 and 51, similar to medea previously, lancer5 casts Gae Bolg
instead of rune of conversion if Q-R is pressed too quickly. Same logic
for Q-W.
For Issue#54,  fix is at line 71 to 75, cooldowns applied explicitly the
runes.