Fix for #33 and #53
For #33 , is-not-hero check is copied from
https://github.com/dun1007/fateanother/blob/master/game/dota_addons/fateanother/scripts/vscripts/addon_game_mode.lua#L1331,
I *think* it should be fine since I have never seen any instance where
something gets killed by no one or some non-hero unit.

For #53 , been unable to replicate it but probably because of 0 latency.
Regardless, for the case where index-2 happens to be the charm, mantra
goes into a3.